tests/wpt/web-platform-tests/trusted-types/HTMLBaseElement-href.tentative.html

<!DOCTYPE html>
<html>
<head>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
  <script src="support/helper.sub.js"></script>

  <script>
    //helper function for the tests
    function testHref(str, url) {
      var base = document.createElement('base');
      base.href = url;
      assert_equals(base.href, str);
    }

    test(t => {
      testHref(URLS.safe, TrustedURL.create(URLS.safe));
    }, "base.href = URLS.safe, TrustedURL.create");

    test(t => {
      testHref(URLS.safe, TrustedURL.unsafelyCreate(URLS.safe));
    }, "base.href = URLS.safe, TrustedURL.unsafelyCreate");
  </script>
</head>
